requirements.txt
fastapi
uvicorn
python-dotenv
pillow
gradio_client
opencv-python
numpy
groq
torch
requests
python-multipart
diffusers
transformers
accelerate